move pliv tests to submodule
