+++ doubleml/iv_model/tests/test_pliv_partial_z_tune.py
@@ -4,17 +4,17 @@ import math
 
 from sklearn.base import clone
 
-from sklearn.linear_model import Lasso
+from sklearn.linear_model import ElasticNet
 
 import doubleml as dml
 
-from ._utils import draw_smpls
-from ..iv_model.tests._utils_pliv_partial_z_manual import fit_pliv_partial_z, boot_pliv_partial_z
+from ...tests._utils import draw_smpls
+from ._utils_pliv_partial_z_manual import fit_pliv_partial_z, boot_pliv_partial_z, tune_nuisance_pliv_partial_z
 
 
 @pytest.fixture(scope='module',
-                params=[Lasso(alpha=0.1)])
-def learner(request):
+                params=[ElasticNet()])
+def learner_r(request):
     return request.param
 
 
@@ -25,13 +25,28 @@ def score(request):
 
 
 @pytest.fixture(scope='module',
-                params=['dml1', 'dml2'])
+                params=['dml2'])
 def dml_procedure(request):
     return request.param
 
 
+@pytest.fixture(scope='module',
+                params=[True, False])
+def tune_on_folds(request):
+    return request.param
+
+
+def get_par_grid(learner):
+    assert learner.__class__ == ElasticNet
+    par_grid = {'l1_ratio': [.1, .5, .7, .9, .95, .99, 1], 'alpha': np.linspace(0.05, 1., 7)}
+    return par_grid
+
+
 @pytest.fixture(scope='module')
-def dml_pliv_partial_z_fixture(generate_data_pliv_partialZ, learner, score, dml_procedure):
+def dml_pliv_partial_z_fixture(generate_data_pliv_partialZ, learner_r, score, dml_procedure, tune_on_folds):
+    par_grid = {'ml_r': get_par_grid(learner_r)}
+    n_folds_tune = 4
+
     boot_methods = ['Bayes', 'normal', 'wild']
     n_folds = 2
     n_rep_boot = 503
@@ -42,7 +57,7 @@ def dml_pliv_partial_z_fixture(generate_data_pliv_partialZ, learner, score, dml_
     z_cols = data.columns[data.columns.str.startswith('Z')].tolist()
 
     # Set machine learning methods for r
-    ml_r = clone(learner)
+    ml_r = clone(learner_r)
 
     np.random.seed(3141)
     obj_dml_data = dml.DoubleMLData(data, 'y', ['d'], x_cols, z_cols)
@@ -50,6 +65,9 @@ def dml_pliv_partial_z_fixture(generate_data_pliv_partialZ, learner, score, dml_
                                               ml_r,
                                               n_folds=n_folds,
                                               dml_procedure=dml_procedure)
+
+    # tune hyperparameters
+    _ = dml_pliv_obj.tune(par_grid, tune_on_folds=tune_on_folds, n_folds_tune=n_folds_tune)
 
     dml_pliv_obj.fit()
 
@@ -60,10 +78,25 @@ def dml_pliv_partial_z_fixture(generate_data_pliv_partialZ, learner, score, dml_
     z = data.loc[:, z_cols].values
     n_obs = len(y)
     all_smpls = draw_smpls(n_obs, n_folds)
+    smpls = all_smpls[0]
+
+    if tune_on_folds:
+        r_params = tune_nuisance_pliv_partial_z(y, x, d, z,
+                                                clone(learner_r),
+                                                smpls, n_folds_tune,
+                                                par_grid['ml_r'])
+    else:
+        xx = [(np.arange(len(y)), np.array([]))]
+        r_params = tune_nuisance_pliv_partial_z(y, x, d, z,
+                                                clone(learner_r),
+                                                xx, n_folds_tune,
+                                                par_grid['ml_r'])
+        r_params = r_params * n_folds
 
     res_manual = fit_pliv_partial_z(y, x, d, z,
-                                    clone(learner),
-                                    all_smpls, dml_procedure, score)
+                                    clone(learner_r),
+                                    all_smpls, dml_procedure, score,
+                                    r_params=r_params)
 
     res_dict = {'coef': dml_pliv_obj.coef,
                 'coef_manual': res_manual['theta'],
